requirements.txt
beautifulsoup4==4.11.1
html5lib==1.1
lxml==4.9.1
pytest==7.1.2
python-dotenv==0.20.0
requests==2.22.0
